compare weights

+++ tools/convert_checkpoint/deepspeed_bloom_to_transformers.py
@@ -50,7 +50,7 @@ if __name__ == "__main__":
         os.makedirs(args.pytorch_dump_folder_path)
 
     config = BloomConfig.from_pretrained(args.bloom_model)
-    config.torch_dtype = torch.bfloat16
+    config.torch_dtype = torch.float16
     config.to_json_file(os.path.join(args.pytorch_dump_folder_path, 'config.json'))
